override newer callbacks in device
